Add error handling
Currently the script is very error prone and does not exit if a command fails.
Instead it will proceed with a possibly bad file state.

This change adds error checks on exit codes where we have seen failures. Furhter
more it adds a wait loop on zplug installation as we have seen that the init.zsh
file is not available right after installation of zplug.